Split algebraic operators in separate files
