+++ include/samurai/petsc/matrix_assembly.hpp
@@ -272,6 +272,12 @@ namespace samurai
         template <class Scheme, class check = void>
         class Assembly : public MatrixAssembly
         {
+            template <class>
+            static constexpr bool dependent_false = false;
+
+            static_assert(
+                dependent_false<typename Scheme::cfg_t>,
+                "Either the required file has not been included, or the Assembly class has not been specialized for this type of scheme.");
         };
 
         template <class Scheme>
